Update index.js
Changed the retrieval of the Debug_Scenes to use the admin object located in this file instead of the user object that I was sending before. This is because there will be a change to how that object works once the user update is finished. The saving of the skill stats also uses the admin object now too.