Update Scene_LoginAndCreate.js
There is now a check to make sure user names and pass's are at a certain length before accepting them.
There was a bug where certain cases of player saving would save the map id as 0 instead of the correct one. This would send the player to a blank map upon logging in. This has been fixed.